Admin login/logout completed
