--- admin/src/App.tsx
@@ -1,10 +1,0 @@
-function App() {
-
-  return (
-    <>
-      Admin Panel
-    </>
-  )
-}
-
-export default App
