--- admin/vite.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  base: "/admin",
-  plugins: [react()],
-})
